+++ lib/daily_games.dart
@@ -74,6 +74,7 @@ void dailyProverb(NyxxGateway client, DotEnv env) async {
     });
   }
 
+  logic();
   cron.schedule(Schedule.parse(frequency), logic);
 }